refactor: PROFILE만 공통화하고 ME_POSTS/ME_COMMENTS/PATCH_USER_PROFILE는 mypage로 롤백
