+++ apps/client/src/shared/api/domain/queries.ts
@@ -1,21 +1,11 @@
-import {
-  infiniteQueryOptions,
-  mutationOptions,
-  queryOptions,
-} from '@tanstack/react-query';
+import { queryOptions } from '@tanstack/react-query';
 import ky from '@toss/ky';
 
-import { USER_MUTATION_KEY, USER_QUERY_KEY } from '@shared/api/keys/query-key';
+import { USER_QUERY_KEY } from '@shared/api/keys/query-key';
 
 import { END_POINT } from '../config/end-point';
 import { api } from '../config/instance';
-import {
-  ImageUploadResponse,
-  MePostResponse,
-  UserProfile,
-  UserProfileEditRequestBody,
-  UserProfileEditResponse,
-} from '../types/types';
+import { ImageUploadResponse, UserProfile } from '../types/types';
 
 // =============================================================================
 // QUERY OPTIONS
@@ -26,24 +16,6 @@ export const USER_QUERY_OPTIONS = {
     queryOptions({
       queryKey: USER_QUERY_KEY.PROFILE(),
       queryFn: getUserProfile,
-    }),
-
-  ME_POSTS: () =>
-    infiniteQueryOptions({
-      queryKey: USER_QUERY_KEY.ME_POSTS(),
-      queryFn: ({ pageParam = 0 }) => getMePosts({ pageParam }),
-      getNextPageParam: (lastPage) =>
-        lastPage.isLast ? undefined : lastPage.nextCursor,
-      initialPageParam: 0,
-    }),
-
-  ME_COMMENTS: () =>
-    infiniteQueryOptions({
-      queryKey: USER_QUERY_KEY.ME_COMMENTS(),
-      queryFn: ({ pageParam = 0 }) => getMeComments({ pageParam }),
-      getNextPageParam: (lastPage) =>
-        lastPage.isLast ? undefined : lastPage.nextCursor,
-      initialPageParam: 0,
     }),
 };
 
@@ -58,49 +30,9 @@ export const getUserProfile = async (): Promise<UserProfile | null> => {
   return response;
 };
 
-export const getMePosts = async ({ pageParam }: { pageParam: number }) => {
-  const url =
-    pageParam === 0
-      ? `${END_POINT.USER.GET_ME_POSTS}?size=10`
-      : `${END_POINT.USER.GET_ME_POSTS}?cursorId=${pageParam}&size=10`;
-
-  const response = await api.get(url).json<MePostResponse>();
-  return response.data;
-};
-
-export const getMeComments = async ({ pageParam }: { pageParam: number }) => {
-  const url =
-    pageParam === 0
-      ? `${END_POINT.USER.GET_ME_COMMENTS}?size=10`
-      : `${END_POINT.USER.GET_ME_COMMENTS}?cursorId=${pageParam}&size=10`;
-
-  const response = await api.get(url).json<MePostResponse>();
-  return response.data;
-};
-
 // =============================================================================
-// MUTATION OPTIONS
+// IMAGE FUNCTIONS
 // =============================================================================
-export const USER_MUTATION_OPTIONS = {
-  PATCH_USER_PROFILE: () => {
-    return mutationOptions({
-      mutationKey: USER_MUTATION_KEY.USER_PROFILE(),
-      mutationFn: ({ body }: { body: UserProfileEditRequestBody }) =>
-        patchUserProfile(body),
-    });
-  },
-};
-
-// =============================================================================
-// MUTATION FUNCTIONS
-// =============================================================================
-
-export const patchUserProfile = async (data: UserProfileEditRequestBody) => {
-  const response = await api
-    .patch(END_POINT.USER.PATCH_USER_INFO, { json: data })
-    .json<UserProfileEditResponse>();
-  return response;
-};
 
 export const postImage = async (mediaType: string[]) => {
   const response = await api
